Add files via upload
This repository include some python projects that were developed as a midterm and final project for a python for data analytics class. The first one (midterm) is a story telling game that logs the input of the user to perform actions in the game. 

The second one a is an application that predicted the winner of the Brazilian championship of 2023, using most pandas, numpy, matplotlib and other libraries available for data analysis.